+++ weldx/core.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any, Dict, List, Tuple, Union
 from warnings import warn
 
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import pint
@@ -15,6 +14,7 @@ from weldx.constants import Q_
 from weldx.time import Time, TimeDependent, types_time_like
 
 if TYPE_CHECKING:
+    import matplotlib.pyplot
     import sympy
 
 __all__ = ["MathematicalExpression", "TimeSeries"]
@@ -290,8 +290,8 @@ class TimeSeries(TimeDependent):
             'step', 'linear'.
 
         """
-        self._data = None
-        self._time_var_name = None
+        self._data: Union[MathematicalExpression, xr.DataArray] = None
+        self._time_var_name = None  # type: str
         self._shape = None
         self._units = None
         self._interp_counter = 0
@@ -326,7 +326,7 @@ class TimeSeries(TimeDependent):
         if not isinstance(self.data, MathematicalExpression):
             if not isinstance(other.data, pint.Quantity):
                 return False
-            return self._data.identical(other.data_array)
+            return self._data.identical(other.data_array)  # type: ignore
 
         return self._data == other.data
 
@@ -616,11 +616,11 @@ class TimeSeries(TimeDependent):
     def plot(
         self,
         time: Union[pd.TimedeltaIndex, pint.Quantity] = None,
-        axes: plt.Axes = None,
+        axes: "matplotlib.pyplot.Axes" = None,
         data_name: str = "values",
         time_unit: Union[str, pint.Unit] = None,
         **mpl_kwargs,
-    ) -> plt.Axes:
+    ) -> "matplotlib.pyplot.Axes":
         """Plot the `TimeSeries`.
 
         Parameters
@@ -644,6 +644,8 @@ class TimeSeries(TimeDependent):
             The matplotlib axes object that was used for the plot
 
         """
+        import matplotlib.pyplot as plt
+
         if axes is None:
             _, axes = plt.subplots()
         if self.is_expression or time is not None:
@@ -655,7 +657,7 @@ class TimeSeries(TimeDependent):
         if time_unit is not None:
             time = time.to(time_unit)
 
-        axes.plot(time.m, self._data.data.m, **mpl_kwargs)
+        axes.plot(time.m, self._data.data.m, **mpl_kwargs)  # type: ignore
         axes.set_xlabel(f"t in {time.u:~}")
         y_unit_label = ""
         if self.units not in ["", "dimensionless"]: