+++ weldx/transformations/local_cs.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import warnings
 from copy import deepcopy
-from typing import TYPE_CHECKING, List, Tuple, Union
+from typing import TYPE_CHECKING, List, Tuple, Union, Any
 
 import numpy as np
 import pandas as pd
@@ -233,8 +233,10 @@ class LocalCoordinateSystem(TimeDependent):
         rhs_cs_inv = rhs_cs.invert()
         return self + rhs_cs_inv
 
-    def __eq__(self: "LocalCoordinateSystem", other: "LocalCoordinateSystem") -> bool:
+    def __eq__(self, other: Any) -> bool:
         """Check equality of LocalCoordinateSystems."""
+        if not isinstance(other, LocalCoordinateSystem):
+            return NotImplemented
 
         def _comp_coords():
             if not isinstance(self.coordinates, type(other.coordinates)):