+++ weldx/util.py
@@ -3,17 +3,19 @@ from __future__ import annotations
 
 import functools
 import json
+import re
 import sys
 import warnings
 from collections.abc import Iterable, Sequence
 from functools import wraps
 from inspect import getmembers, isfunction
 from pathlib import Path
-from typing import Any, Callable, Collection, Dict, List, Mapping, Union
+from typing import Any, Callable, ClassVar, Collection, Dict, List, Mapping, Union
 
 import numpy as np
 import pandas as pd
 import pint
+import psutil
 import xarray as xr
 from asdf.tags.core import NDArrayType
 from boltons import iterutils
@@ -333,7 +335,7 @@ def matrix_is_close(mat_a, mat_b, abs_tol=1e-9) -> bool:
 
     if mat_a.shape != mat_b.shape:
         return False
-    return np.all(np.isclose(mat_a, mat_b, atol=abs_tol))
+    return np.all(np.isclose(mat_a, mat_b, atol=abs_tol)).__bool__()
 
 
 def vector_is_close(vec_a, vec_b, abs_tol=1e-9) -> bool:
@@ -359,7 +361,7 @@ def vector_is_close(vec_a, vec_b, abs_tol=1e-9) -> bool:
 
     if vec_a.size != vec_b.size:
         return False
-    return np.all(np.isclose(vec_a, vec_b, atol=abs_tol))
+    return np.all(np.isclose(vec_a, vec_b, atol=abs_tol)).__bool__()
 
 
 def mat_vec_mul(a, b) -> np.ndarray:
@@ -482,7 +484,7 @@ def xr_matmul(
 
     mul_func = np.matmul
     if len(dims_a) > len(dims_b):
-        mul_func = mat_vec_mul
+        mul_func = mat_vec_mul  # type: ignore[assignment] # irrelevant for us
 
     if trans_a:
         dims_a = reversed(dims_a)
@@ -683,7 +685,7 @@ def xr_interp_like(
     return result
 
 
-def _check_dtype(var_dtype, ref_dtype: dict) -> bool:
+def _check_dtype(var_dtype, ref_dtype: str) -> bool:
     """Check if dtype matches a reference dtype (or is subdtype).
 
     Parameters
@@ -1139,12 +1141,12 @@ class _Eq_compare_nested:
     """Compares nested data structures like lists, sets, tuples, arrays, etc."""
 
     # some types need special comparison handling.
-    compare_funcs = {
+    compare_funcs: ClassVar = {
         (np.ndarray, NDArrayType, pint.Quantity, pd.Index): _array_equal,
         (xr.DataArray, xr.Dataset): lambda x, y: x.identical(y),
     }
     # these types will be treated as equivalent.
-    _type_equalities = [
+    _type_equalities: ClassVar = [
         (np.ndarray, NDArrayType),
     ]
 
@@ -1249,7 +1251,7 @@ compare_nested = _Eq_compare_nested.compare_nested
 def is_interactive_session() -> bool:
     """Check whether this Python session is interactive, e.g. Jupyter/IPython."""
     try:
-        get_ipython = sys.modules["IPython"].get_ipython
+        get_ipython = sys.modules["IPython"].get_ipython  # type: ignore[attr-defined]
         if not get_ipython():
             return False
         if "IPKernelApp" not in get_ipython().config:  # pragma: no cover
@@ -1258,3 +1260,14 @@ def is_interactive_session() -> bool:
         return False
     else:
         return True
+
+
+def is_jupyterlab_session() -> bool:
+    """Heuristic to check whether we are in a Jupyter-Lab session.
+
+    Notes
+    -----
+    False positive, if classic NB launched from JupyterLab.
+
+    """
+    return any(re.search("jupyter-lab", x) for x in psutil.Process().parent().cmdline())